removed fstrings, moved util methods to utils, code cleanup
